[ADD] Modify Command 추가

+++ src/main/java/com/samsungds/codereview/teamd/command/DelCommand.java
@@ -18,19 +18,16 @@ public class DelCommand implements ICommand{
     @Override
     public Boolean execute(String inputStr) throws IOException {
         if(irepo == null) throw new NullPointerException("Error : Repository Link");
-        System.out.println("Repository Null Check OK");
 
         ArrayList<String> itemList = inputStringToArrayList(inputStr);
-        System.out.println("itemList Creation OK");
 
         if(!(itemList.get(Constants.INPUT_STR_COMMAND_POS).equals(Constants.COMMAND_DEL)))
             throw new IllegalArgumentException();
-        System.out.println("Input String Check OK");
+
+        if(itemList.size() != 6) throw new IllegalArgumentException("Error : Argument Count");
 
         Map<Integer, Employee> map = irepo.delete(itemList.get(Constants.INPUT_STR_KEY1),
                 itemList.get(Constants.INPUT_STR_VALUE1));
-
-        System.out.println("Get Map Data OK");
 
         ArrayList<Employee> empList = new ArrayList<>();
 
@@ -40,13 +37,7 @@ public class DelCommand implements ICommand{
             }
         }
 
-        System.out.println("Make Employee List OK");
-
         printResult(empList, isPrintOptionEnable(itemList.get(Constants.INPUT_STR_OPTION1_POS)));
-
-        System.out.println("Print Result");
-        System.out.println("COMMAND : " + Constants.COMMAND_DEL);
-        System.out.println("Total Cnt : " + empList.size());
 
         return true;
     }
@@ -75,9 +66,4 @@ public class DelCommand implements ICommand{
     private Boolean isPrintOptionEnable(String inputStr){
         return Constants.OPTION1_PRINT.equals(inputStr);
     }
-
-//    private ArrayList<Employee> makeEmployeeList(Map<Integer, Employee> map){
-//        return null;
-//    }
-
 }
